Zerodev/kernel v2

+++ src/tests/Incrementer.sol
@@ -3,7 +3,7 @@
 pragma solidity ^0.8.18;
 
 contract Incrementer {
-    uint public value = 0;
+    uint256 public value = 0;
 
     function increment() external {
         value = value + 1;
